+++ client/package.json
@@ -3,9 +3,9 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "next dev",
+    "dev": "next dev -p 3001",
     "build": "next build && next export",
-    "start": "next start",
+    "start": "next start -p 3001",
     "lint": "next lint",
     "deploy": "next build && next export && firebase deploy --only hosting"
   },
