+++ client/pages/dashboard.tsx
@@ -53,7 +53,7 @@ const DashboardPage = () => {
     }
   }, [theme]);
 
-  const domainUrl = "http://localhost:3001/";
+  const domainUrl = "http://localhost:3000/";
 
   return isLoading ? (
     <LoadingScreen />
