+++ server/fly.toml
@@ -1,0 +1,17 @@
+# fly.toml app configuration file generated for linkzar on 2023-08-30T16:57:53+05:00
+#
+# See https://fly.io/docs/reference/configuration/ for information about how to use this file.
+#
+
+app = "linkzar"
+primary_region = "lax"
+
+[build]
+
+[http_service]
+  internal_port = 3000
+  force_https = true
+  auto_stop_machines = false
+  auto_start_machines = true
+  min_machines_running = 0
+  processes = ["app"]
